Quantnuis-Backend/requirements.txt
# ==============================================================================
# QUANTNUIS - DÉPENDANCES
# ==============================================================================
# Versions compatibles avec Python 3.11 et AWS Lambda
# ==============================================================================

# --- CŒUR SCIENTIFIQUE ---
# NumPy < 2.0 est OBLIGATOIRE pour TensorFlow 2.15
numpy<2.0.0

# Pandas & Scikit-learn (versions stables avec numpy<2)
pandas==2.1.4
scikit-learn==1.4.2

# --- AUDIO & IA ---
# Librosa 0.9.2 est la version la plus stable sur AWS Lambda
# Évite la compilation de soxr qui pose des problèmes
librosa==0.9.2

# Dépendances audio
soundfile>=0.12.0
pooch>=1.0
resampy>=0.4.0
pydub>=0.25.0

# --- TENSORFLOW ---
# TF 2.15 est la dernière version compatible avec l'image AWS Lambda standard
# Supporte Python 3.9 à 3.11 (PAS 3.12 !)
tensorflow==2.15.0
keras<3.0.0
h5py<3.11.0

# --- DATA AUGMENTATION ---
imbalanced-learn==0.12.4

# --- VISUALISATION ---
matplotlib>=3.7.5
seaborn>=0.13.0

# --- UTILITAIRES ---
joblib>=1.3.0
scipy<1.13.0

# --- SERVEUR WEB ---
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# --- AWS ---
mangum>=0.17.0
boto3>=1.34.0

# --- BASE DE DONNÉES ---
sqlalchemy>=2.0.0

# --- SÉCURITÉ ---
# bcrypt 4.0.x est compatible avec passlib 1.7.4
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0

# --- VALIDATION ---
pydantic>=2.0.0
email-validator>=2.0.0
